# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ICAMMDA Website API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /team:
    get:
      operationId: listTeamMembers
      tags: [team]
      summary: List all team members
      responses:
        "200":
          description: List of team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: createTeamMember
      tags: [team]
      summary: Create a team member
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamMemberBody"
      responses:
        "201":
          description: Created team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /team/{id}:
    patch:
      operationId: updateTeamMember
      tags: [team]
      summary: Update a team member
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamMemberBody"
      responses:
        "200":
          description: Updated team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
        "404":
          description: Not found
    delete:
      operationId: deleteTeamMember
      tags: [team]
      summary: Delete a team member
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /research-groups:
    get:
      operationId: listResearchGroups
      tags: [research]
      summary: List all research groups with their members
      responses:
        "200":
          description: List of research groups with members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResearchGroupWithMembers"
    post:
      operationId: createResearchGroup
      tags: [research]
      summary: Create a research group
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResearchGroupBody"
      responses:
        "201":
          description: Created research group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchGroup"
  /research-groups/{id}:
    patch:
      operationId: updateResearchGroup
      tags: [research]
      summary: Update a research group
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResearchGroupBody"
      responses:
        "200":
          description: Updated research group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchGroup"
        "404":
          description: Not found
    delete:
      operationId: deleteResearchGroup
      tags: [research]
      summary: Delete a research group
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /research-members:
    get:
      operationId: listResearchMembers
      tags: [research]
      summary: List research members (optionally filter visiting scholars)
      parameters:
        - name: visiting
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of research members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResearchMember"
    post:
      operationId: createResearchMember
      tags: [research]
      summary: Create a research member
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResearchMemberBody"
      responses:
        "201":
          description: Created research member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchMember"
  /research-members/{id}:
    patch:
      operationId: updateResearchMember
      tags: [research]
      summary: Update a research member
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResearchMemberBody"
      responses:
        "200":
          description: Updated research member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchMember"
        "404":
          description: Not found
    delete:
      operationId: deleteResearchMember
      tags: [research]
      summary: Delete a research member
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /research-publications:
    get:
      operationId: listResearchPublications
      tags: [research]
      summary: List research publications
      responses:
        "200":
          description: List of publications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResearchPublication"
    post:
      operationId: createResearchPublication
      tags: [research]
      summary: Create a research publication
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResearchPublicationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchPublication"
  /research-publications/{id}:
    patch:
      operationId: updateResearchPublication
      tags: [research]
      summary: Update a publication
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResearchPublicationBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchPublication"
        "404":
          description: Not found
    delete:
      operationId: deleteResearchPublication
      tags: [research]
      summary: Delete a publication
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    TeamMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        title:
          type: string
        bio:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        role:
          type: string
          description: "director | researcher | staff | postdoc"
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - title
        - role
        - order
        - createdAt
        - updatedAt
    CreateTeamMemberBody:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        bio:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        role:
          type: string
        order:
          type: integer
      required:
        - name
        - title
        - role
    UpdateTeamMemberBody:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        bio:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        role:
          type: string
        order:
          type: integer
...
    ResearchGroup:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - order
        - createdAt
        - updatedAt
    ResearchMember:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: ["integer", "null"]
        name:
          type: string
        role:
          type: string
        affiliation:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        isVisiting:
          type: boolean
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - role
        - isVisiting
        - order
        - createdAt
        - updatedAt
    ResearchGroupWithMembers:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        members:
          type: array
          items:
            $ref: "#/components/schemas/ResearchMember"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - order
        - members
        - createdAt
        - updatedAt
    CreateResearchGroupBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
      required:
        - name
    UpdateResearchGroupBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
    CreateResearchMemberBody:
      type: object
      properties:
        groupId:
          type: ["integer", "null"]
        name:
          type: string
        role:
          type: string
        affiliation:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        isVisiting:
          type: boolean
        order:
          type: integer
      required:
        - name
        - role
    UpdateResearchMemberBody:
      type: object
      properties:
        groupId:
          type: ["integer", "null"]
        name:
          type: string
        role:
          type: string
        affiliation:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        isVisiting:
          type: boolean
        order:
          type: integer
    ResearchPublication:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        authors:
          type: string
        journal:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        doi:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        abstract:
          type: ["string", "null"]
        displayOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - authors
        - displayOrder
        - createdAt
        - updatedAt
    CreateResearchPublicationBody:
      type: object
      properties:
        title:
          type: string
        authors:
          type: string
        journal:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        doi:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        abstract:
          type: ["string", "null"]
        displayOrder:
          type: integer
      required:
        - title
        - authors
    UpdateResearchPublicationBody:
      type: object
      properties:
        title:
          type: string
        authors:
          type: string
        journal:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        doi:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        abstract:
          type: ["string", "null"]
        displayOrder:
          type: integer
...